make LDAP code use DI everywhere
